requirements.txt
appnope==0.1.3
asn1crypto==1.5.1
asttokens==2.2.1
backcall==0.2.0
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
importlib-metadata==6.7.0
ipykernel==6.23.3
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pg8000==1.29.8
pickleshare==0.7.5
platformdirs==3.8.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydot==1.4.2
Pygments==2.15.1
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pyzmq==25.1.0
scramp==1.4.4
six==1.16.0
SQLAlchemy==2.0.17
sqlalchemy-schemadisplay==1.3
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.3
wcwidth==0.2.6
zipp==3.15.0
faker==18.11.1
psycopg2-binary==2.9.6